pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-investor-brief"
version = "0.1.0"
description = "AI agent that creates financial briefs for investors"
readme = "README.md"
authors = [
    {name = "Lina Faik", email = "linafaik08@gmail.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"

# Core dependencies - minimal and focused
dependencies = [
    # MLflow and AI framework
    "mlflow>=3.0.0",
    "openai>=1.109",
    
    # Financial data and web scraping
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Configuration and utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    "tavily-python>=0.7.12"
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

# Enhanced features (optional)
enhanced = [
    "tavily-python>=0.3.0",  # Professional web search
    "alpha-vantage>=2.3.0",  # Advanced financial APIs
    "newspaper3k>=0.2.8",    # News article extraction
    "plotly>=5.15.0",        # Interactive visualizations
    "streamlit>=1.28.0",     # Web UI (future enhancement)
]

# All optional dependencies
all = [
    "agent-investor-brief[dev,enhanced]"
]

[project.urls]
Homepage = "https://github.com/yourusername/agent-investor-brief"
Repository = "https://github.com/yourusername/agent-investor-brief"
Issues = "https://github.com/yourusername/agent-investor-brief/issues"

[project.scripts]
# CLI commands for the package
investor-brief = "agent_investor_brief.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_investor_brief"]

[tool.hatch.envs.default]
dependencies = [
    "agent-investor-brief[dev]"
]

[tool.hatch.envs.default.scripts]
# Convenient development scripts
test = "pytest {args:tests}"
test-cov = "pytest --cov=agent_investor_brief {args:tests}"
cov-report = "pytest --cov=agent_investor_brief --cov-report=html {args:tests}"
lint = "flake8 src tests"
format = "black src tests && isort src tests"
type-check = "mypy src"
notebook = "jupyter notebook notebooks/"
mlflow-ui = "mlflow ui"

# Setup and demo commands
setup = "python -c 'from src.agent_investor_brief.config import setup_environment; setup_environment()'"
demo = "jupyter notebook notebooks/demo.ipynb"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agent_investor_brief"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "mlflow.*",
    "langchain.*",
    "beautifulsoup4.*",
    "requests.*",
]
ignore_missing_imports = true

# UV-specific configuration
[tool.uv]
# UV workspace configuration
dev-dependencies = [
    "agent-investor-brief[dev]"
]

# UV will use this for faster resolution
index-url = "https://pypi.org/simple"